Finalise transfer of all methods of proofs.go to new package
